+++ MultiplayerModLimit/docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.0
 
 - 添加`SendSMAPIInfo`选项，让玩家可以选择是否向不满足模组要求的客机玩家发送SMAPI信息
+- 添加`add_allow`、`add_require`和`add_ban`命令，使玩家能方便的向当前模组列表中添加模组
 
 # 0.2.2